+++ app/components/ChatMessage/style.tailwind.ts
@@ -1,0 +1,2 @@
+export const userStyles = "bg-gray-700 text-white rounded-lg p-3 max-w-xs self-end";
+export const botStyles = "bg-blue-800 text-white rounded-lg p-3 max-w-xs self-start";
