+++ app/components/ChatMessage/type.ts
@@ -1,0 +1,5 @@
+
+export type ChatMessageProps = {
+    content: string;
+    role: string;
+  }
