+++ app/layout.tsx
@@ -4,8 +4,8 @@ import "./globals.css";
 
 
 export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+  title: "A79 - Explore Ag Grid",
+  description: "Explore Ag Grid with next app",
 };
 
 export default function RootLayout({
